+++ packages/adminui-vue/components/controls/UFile/views/FileRenderer.vue
@@ -77,9 +77,9 @@ const CONTENT_TYPE_ICONS = {
   'text/html': 'u-icon-file-html',
   'text/xml': 'u-icon-file-html',
   'text/markdown': 'u-icon-file-html',
-  'text/javascript': 'u-icon-file-html',
+  'text/javascript': 'fas fa-cogs',
   'text/css': 'u-icon-file-html',
-  'text/cmd': 'u-icon-file-html',
+  'text/cmd': 'fas fa-cogs',
 
   'text/csv': 'u-icon-file-csv',
 
@@ -103,6 +103,7 @@ const CONTENT_TYPE_ICONS = {
   'application/x-rar-compressed': 'fa fa-file-archive-o',
   'application/x-7z-compressed': 'fa fa-file-archive-o',
   'application/zip': 'fa fa-file-archive-o',
+  'application/gzip': 'fa fa-file-archive-o'
 }
 
 export default {
@@ -132,9 +133,11 @@ export default {
 
   computed: {
     icon () {
-      if (CONTENT_TYPE_ICONS[this.file.ct]) {
-        return CONTENT_TYPE_ICONS[this.file.ct]
-      } else if (this.file.ct.startsWith('video/')) {
+      if (!this.file.ct) return 'u-icon-file'
+      const ct = this.file.ct.split(';')[0] || '' // `text/plain; charset=utf8` -> text/plain
+      if (CONTENT_TYPE_ICONS[ct]) {
+        return CONTENT_TYPE_ICONS[ct]
+      } else if (ct.startsWith('video/')) {
         return 'far fa-file-video'
       } else {
         return 'u-icon-file'